Clear cached local metadata during one test
This is a test that changes the value of the "CONDA_USE_ONLY_TAR_BZ2"
environment variable.  That affects the metadata that is returned, so
the cache must be cleared when it changes.   This issue does not arise
in production as the cache is in memory and the variable doesn't
change once the conda process is up and running.